refactor(app): extract question-dock pure logic to question-tool-respond
Move resolveSkipAction, normalizeToolRespondError, createQuestionResponseGuard
(plus their private helpers and types) out of session-question-dock.tsx into a
new question-tool-respond.ts module. Pure extraction, no behavior change;
isSameQuestionRequest stays in the dock since QuestionRequestFingerprint is
shared by the component body.
